+++ stwo_cairo_prover/crates/prover/src/cairo_air/prover.rs
@@ -249,8 +249,8 @@ pub mod tests {
 
         use crate::cairo_air::debug_tools::assert_constraints::assert_cairo_constraints;
         use crate::cairo_air::preprocessed::PreProcessedTrace;
+        use crate::cairo_air::prover::prove_cairo;
         use crate::cairo_air::prover::tests::{test_basic_cairo_air_input, test_cfg};
-        use crate::cairo_air::prover::{prove_cairo, ProverInput};
         use crate::cairo_air::verifier::verify_cairo;
 
         // TODO(Ohad): fine-grained constraints tests.
@@ -308,111 +308,6 @@ pub mod tests {
                 .collect_vec();
 
             assert!(proofs.iter().all_equal());
-        }
-
-        /// These tests' inputs were generated using cairo-vm with 50 instances of each builtin.
-        pub mod builtin_tests {
-            use std::fs::File;
-
-            use cairo_vm::air_public_input::MemorySegmentAddresses;
-            use stwo_cairo_adapter::memory::MemoryEntryIter;
-            use stwo_cairo_adapter::vm_import::generate_test_input;
-
-            use super::*;
-
-            /// Asserts that all builtins are present in the input.
-            /// Panics if any of the builtins is missing.
-            fn assert_all_builtins_in_input(input: &ProverInput) {
-                let empty_builtins = input
-                    .builtins_segments
-                    .get_counts()
-                    .iter()
-                    .filter(|(_, &count)| count == 0)
-                    .map(|(name, _)| format!("{:?}", name))
-                    .collect_vec();
-                assert!(
-                    empty_builtins.is_empty(),
-                    "The following builtins are missing: {}",
-                    empty_builtins.join(", ")
-                );
-            }
-
-            #[test]
-            fn test_prove_verify_all_builtins() {
-                let input = generate_test_input("test_prove_verify_all_builtins");
-                assert_all_builtins_in_input(&input);
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
-
-            #[test]
-            fn test_prove_verify_add_mod_builtin() {
-                let input = generate_test_input("test_prove_verify_add_mod_builtin");
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
-
-            /// Asserts that there is an unused `add` value in the first instance in bitwise
-            /// builtin segment, inducing a "hole".
-            fn assert_bitwise_builtin_has_holes(
-                test_name: &str,
-                bitwise_segment: &Option<MemorySegmentAddresses>,
-            ) {
-                let bitwise_segment = bitwise_segment.as_ref().unwrap();
-                let mut d = std::path::PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-                d.push("../../test_data/");
-                d.push(test_name);
-                let mut memory_file =
-                    std::io::BufReader::new(File::open(d.join("mem").as_path()).unwrap());
-                let memory_entries = MemoryEntryIter(&mut memory_file).collect_vec();
-                assert!(memory_entries
-                    .iter()
-                    .all(|entry| entry.address != (bitwise_segment.begin_addr + 2) as u64));
-            }
-
-            #[test]
-            fn test_prove_verify_bitwise_builtin() {
-                let input = generate_test_input("test_prove_verify_bitwise_builtin");
-                assert_bitwise_builtin_has_holes(
-                    "test_prove_verify_bitwise_builtin",
-                    &input.builtins_segments.bitwise,
-                );
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
-
-            #[test]
-            fn test_prove_verify_mul_mod_builtin() {
-                let input = generate_test_input("test_prove_verify_mul_mod_builtin");
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
-
-            #[test]
-            fn test_prove_verify_range_check_bits_96_builtin() {
-                let input = generate_test_input("test_prove_verify_range_check_bits_96_builtin");
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
-
-            #[test]
-            fn test_prove_verify_range_check_bits_128_builtin() {
-                let input = generate_test_input("test_prove_verify_range_check_bits_128_builtin");
-                let cairo_proof =
-                    prove_cairo::<Blake2sMerkleChannel>(input, test_cfg(), PcsConfig::default())
-                        .unwrap();
-                verify_cairo::<Blake2sMerkleChannel>(cairo_proof, PcsConfig::default()).unwrap();
-            }
         }
     }
 }
